Unify jitbuilder/compilertest/compiler IlGeneratorMethodDetails
The two compiler extensions jitbuilder and compilertest both define
virtually identical IlGeneratorMethodDetails classes. Rather than
force compiler extensions to copy code from one of these extensions
or to write all that code themselves, the OMR compiler component
could simply provide these facilities in OMR::IlGeneratorMethodDetails.
Since the current classes are all based on TR_ResolvedMethod,
there isn't anything special about them and can just move into the
OMR compiler. At the same time, some improvements were made to
make more methods const that shouldn't change the underlying object.

This removes 3 files each from jitbuilder and fvtest/compilertest
and eliminates the JitBuilder ilgen directory entirely.
