+++ compiler/ilgen/CMakeLists.txt
@@ -22,6 +22,7 @@
 compiler_library(ilgen
 	${CMAKE_CURRENT_LIST_DIR}/IlGenRequest.cpp
 	${CMAKE_CURRENT_LIST_DIR}/IlInjector.cpp
+	${CMAKE_CURRENT_LIST_DIR}/OMRIlGeneratorMethodDetails.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRBytecodeBuilder.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRIlBuilder.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRIlType.cpp
