+++ compiler/ilgen/OMRIlGeneratorMethodDetails.cpp
@@ -31,28 +31,25 @@
 #include "env/IO.hpp"
 #include "ras/Logger.hpp"
 
-namespace JitBuilder {
-
-IlGeneratorMethodDetails::IlGeneratorMethodDetails(TR_ResolvedMethod *method)
-    : OMR::IlGeneratorMethodDetailsConnector()
-    , _method(static_cast<TR::ResolvedMethod *>(method))
+OMR::IlGeneratorMethodDetails::IlGeneratorMethodDetails(TR_ResolvedMethod *method)
+    : _method(static_cast<TR::ResolvedMethod *>(method))
+    , _ilVerifier(NULL)
 {}
 
-bool IlGeneratorMethodDetails::sameAs(TR::IlGeneratorMethodDetails &other, TR_FrontEnd *fe)
+bool OMR::IlGeneratorMethodDetails::sameAs(TR::IlGeneratorMethodDetails &other, TR_FrontEnd *) const
 {
     return self()->getMethod() == other.getMethod();
 }
 
-TR_IlGenerator *IlGeneratorMethodDetails::getIlGenerator(TR::ResolvedMethodSymbol *methodSymbol, TR_FrontEnd *fe,
+TR_IlGenerator *OMR::IlGeneratorMethodDetails::getIlGenerator(TR::ResolvedMethodSymbol *methodSymbol, TR_FrontEnd *fe,
     TR::Compilation *comp, TR::SymbolReferenceTable *symRefTab, bool forceClassLookahead,
     TR_InlineBlocks *blocksToInline)
 {
     TR_ASSERT(forceClassLookahead == false, "IlGenerator does not support class lookahead");
     TR_ASSERT(blocksToInline == 0, "IlGenerator does not yet support partial inlining");
     TR::ResolvedMethod *method = static_cast<TR::ResolvedMethod *>(methodSymbol->getResolvedMethod());
-    return (TR_IlGenerator *)method->getInjector(self(), methodSymbol, static_cast<TR::FrontEnd *>(fe), symRefTab);
+    TR::FrontEnd *fe = static_cast<TR::FrontEnd *>(trfe);
+    return method->getIlGenerator(self(), methodSymbol, fe, symRefTab);
 }
 
-void IlGeneratorMethodDetails::print(OMR::Logger *log, TR_FrontEnd *fe) { log->printf("( %p )", self()->getMethod()); }
-
-} // namespace JitBuilder
+void OMR::IlGeneratorMethodDetails::print(OMR::Logger *log, TR_FrontEnd *fe) { log->printf("( %p )", self()->getMethod()); }
