+++ compiler/ilgen/OMRIlGeneratorMethodDetails.hpp
@@ -39,12 +39,18 @@ typedef OMR::IlGeneratorMethodDetails IlGeneratorMethodDetailsConnector;
 #include "infra/Annotations.hpp"
 
 class TR_FrontEnd;
+class TR_IlGenerator;
+class TR_IlVerifier;
+class TR_InlineBlocks;
 class TR_ResolvedMethod;
-
 namespace TR {
+class Compilation;
 class IlGeneratorMethodDetails;
 class IlVerifier;
-} // namespace TR
+class ResolvedMethod;
+class ResolvedMethodSymbol;
+class SymbolReferenceTable;
+}
 
 namespace OMR {
 
@@ -57,23 +63,31 @@ class Logger;
  */
 class OMR_EXTENSIBLE IlGeneratorMethodDetails {
 public:
+    IlGeneratorMethodDetails(TR_ResolvedMethod *method);
+
     virtual bool isMethodInProgress() const { return false; }
 
     bool supportsInvalidation() { return false; }
 
-    bool sameAs(TR::IlGeneratorMethodDetails &other) { return false; }
+    TR::ResolvedMethod * getMethod() const { return _method; }
+    TR_ResolvedMethod * getResolvedMethod() const { return (TR_ResolvedMethod *) _method; }
 
-    void print(OMR::Logger *log, TR_FrontEnd *fe) {}
+    bool sameAs(TR::IlGeneratorMethodDetails & other, TR_FrontEnd *) const;
+
+    void print(OMR::Logger *log, TR_FrontEnd *fe);
 
     inline static TR::IlGeneratorMethodDetails &create(TR::IlGeneratorMethodDetails &target, TR_ResolvedMethod *method);
 
-    TR::IlVerifier *getIlVerifier() { return _ilVerifier; }
+    void setIlVerifier(TR::IlVerifier * ilVerifier)      { _ilVerifier = ilVerifier; }
+    TR::IlVerifier * getIlVerifier()                     { return _ilVerifier; }
 
-    void setIlVerifier(TR::IlVerifier *ilVerifier) { _ilVerifier = ilVerifier; }
+    TR_IlGenerator *getIlGenerator(TR::ResolvedMethodSymbol *methodSymbol, TR_FrontEnd * fe, TR::Compilation *comp,
+        TR::SymbolReferenceTable *symRefTab, bool forceClassLookahead, TR_InlineBlocks *blocksToInline);
 
 protected:
     IlGeneratorMethodDetails()
-        : _ilVerifier(NULL)
+        : _method(NULL)
+        , _ilVerifier(NULL)
     {}
 
     inline TR::IlGeneratorMethodDetails *self();
@@ -82,7 +96,6 @@ protected:
     virtual ~IlGeneratorMethodDetails() {}
 
     void *operator new(size_t size, TR::IlGeneratorMethodDetails *p) { return (void *)p; }
-
     void *operator new(size_t size, TR::IlGeneratorMethodDetails &p) { return (void *)&p; }
 
     void operator delete(void *pMem, TR::IlGeneratorMethodDetails *p) {};
@@ -90,6 +103,7 @@ protected:
 
     void operator delete(void *pMem, size_t size) { ::operator delete(pMem); };
 
+    TR::ResolvedMethod * _method;
     TR::IlVerifier *_ilVerifier;
 };
